Changed type of i

+++ 0x19-hash_tables/0-hash_table_create.c
@@ -7,7 +7,7 @@
 hash_table_t *hash_table_create(unsigned long int size)
 {
 	hash_table_t *new_table;
-	unsigned long int i;
+	unsigned int i;
 
 	if (size < 1)
 		return (NULL);
